表示画面をコンポーネントにした
