+++ package.json
@@ -18,6 +18,7 @@
     "next": "13.1.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "sass": "^1.58.1",
     "typescript": "4.9.5"
   }
 }
